requirements.txt
appdirs==1.4.4
asgiref==3.2.7
atomicwrites==1.4.0
attrs==20.3.0
cassandra-driver==3.23.0
certifi==2020.4.5.2
chardet==3.0.4
click==7.1.2
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cql==1.4.0
cqlsh==5.0.4
distlib==0.3.1
dj-database-url==0.5.0
Django==3.0.7
djangorestframework==3.11.0
drf-nested-resources==2.0.0
filelock==3.0.12
geomet==0.1.2
greenlet==1.0.0
gunicorn==20.0.4
idna==2.9
importlib-metadata==3.4.0
importlib-resources==5.1.2
iniconfig==1.1.1
itypes==1.2.0
Jinja2==2.11.2
jsonschema==3.2.0
lxml==4.5.0
Markdown==3.2.1
MarkupSafe==1.1.1
packaging==20.9
pipenv==2020.11.15
pluggy==0.13.1
psycopg2-binary==2.8.4
py==1.10.0
pyang==2.2.1
Pygments==2.6.1
pyparsing==2.4.7
pyrecord==1.0.1
pyrsistent==0.17.3
pytest==6.2.2
pytz==2020.1
PyYAML==5.3.1
requests==2.23.0
six==1.15.0
SQLAlchemy==1.4.2
sqlparse==0.3.1
thrift==0.13.0
toml==0.10.2
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.25.9
virtualenv==20.4.2
virtualenv-clone==0.5.4
whitenoise==5.0.1
zipp==3.4.0
